+++ src/mddj/_cli/read/name.py
@@ -1,0 +1,10 @@
+import click
+
+from mddj._cli.state import CommandState, common_args
+
+
+@click.command("name")
+@common_args
+def read_name(*, state: CommandState) -> None:
+    """Read the 'Name' of the current project."""
+    click.echo(state.dj.read.name())