+++ changelog.rst
@@ -9,6 +9,7 @@ Changelog
 - Add ``DJ.read.import_names()`` and ``mddj read import-names``
 - Add ``DJ.read.import_namespaces()`` and ``mddj read import-namespaces``
 - Add ``DJ.read.optional_dependencies()`` and ``mddj read optional-dependencies``
+- Add ``DJ.read.authors()`` and ``mddj read authors``
 
 0.4.2
 -----