+++ changelog.rst
@@ -4,12 +4,16 @@ Changelog
 .. Unreleased
 .. ----------
 
+0.5.0
+-----
+
 - Add ``DJ.read.name()`` and ``mddj read name``
 - Add ``DJ.read.classifiers()`` and ``mddj read classifiers``
 - Add ``DJ.read.import_names()`` and ``mddj read import-names``
 - Add ``DJ.read.import_namespaces()`` and ``mddj read import-namespaces``
 - Add ``DJ.read.optional_dependencies()`` and ``mddj read optional-dependencies``
 - Add ``DJ.read.authors()`` and ``mddj read authors``
+- Add ``DJ.read.maintainers()`` and ``mddj read maintainers``
 
 0.4.2
 -----